refactor, move example questions elsewhere
