+++ src/components/chat.tsx
@@ -15,14 +15,9 @@ import { useChat } from "ai/react";
 import { Grid } from "react-loader-spinner";
 import Bubble from "./chat/bubble";
 import { welcomeMessage } from "@/lib/strings";
+import ExampleQuestions from "./chat/example-questions";
 
 export default function Chat() {
-  const examples = [
-    "What hotels are there in Hawana Salalah?",
-    "What is there to do in Hawana Salalah?",
-    "How much can I expect to pay staying in Hawana Salalah?",
-  ];
-
   const {
     messages,
     input,
@@ -68,22 +63,11 @@ export default function Chat() {
               id: "initialai",
             }}
           />
-          { messages.length === 0 && (
-            <div className="flex flex-col space-y-4 border-t border-gray-200 bg-gray-50 p-7 sm:p-10">
-              {examples.map((example, i) => (
-                <button
-                  key={i}
-                  className="rounded-md border border-gray-200 bg-white px-5 py-3 text-left text-sm text-gray-500 transition-all duration-75 hover:border-black hover:text-gray-700 active:bg-gray-50"
-                  onClick={() => {
-                    setInput(example);
-                    inputRef.current?.focus();
-                  }}
-                >
-                  {example}
-                </button>
-              ))}
-            </div>
-          )}
+          <ExampleQuestions
+            messages={messages}
+            setInput={setInput}
+            inputRef={inputRef}
+          />
           {messages.map((message) => (
             <Bubble key={message.id} message={message} />
           ))}
